Add Room tables for activities and actions
Activity represents the content in the phone number/username input field, the source by which the content was received (whether by clicking a link, sharing with the app etc), and the time at which the content was received.
Action represents the user phone number selection (if multiple phone numbers were found in the input field), and the messenger app that was launched by clicking a button.