Add PhoneNumberHelper method to convert local number to international format
This adds a new method that accepts a local or international-format phone number and a country code and returns the phone number in international format.
This is intended to be used when user input is formatted with a country code based on the default country code or most recently used country code.