Add extension for conditional data migration
The conditional data migration accepts a boolean-returning condition that determines whether the data migration is to be run.
If the condition is true, it performs the action. It is the responsibility of the developer to define a condition such that it would not be true in multiple executions if the action is to be executed once and only once.